# CilAssembler diagnostics performance regression