… test/AsmResolver.Benchmarks/PEReadWriteBenchmark.cs
… test/AsmResolver.Benchmarks/Resources/HelloWorld.ManyMethods.deflate