Adds constructor to `CilAssembler` that accepts a delegate for lazy loading the diagnostic prefix, in order to avoid expensive string formatting of method full names.

Closes #260